+++ llama_index/llms/__init__.py
@@ -22,6 +22,7 @@ from llama_index.llms.palm import PaLM
 from llama_index.llms.predibase import PredibaseLLM
 from llama_index.llms.replicate import Replicate
 from llama_index.llms.xinference import Xinference
+from llama_index.llms.portkey import Portkey
 
 __all__ = [
     "OpenAI",
@@ -45,5 +46,6 @@ __all__ = [
     "CompletionResponseAsyncGen",
     "LLMMetadata",
     "Xinference",
+    "Portkey",
     "MonsterLLM",
 ]